refactor(web): Tool UI を純コンポーネントと helper に分離してテストを追加する
ChartToolUI から Chart 本体と chart-helpers (getColorAt /
formatPiePercent) を切り出し、ChoiceButtonsToolUI から
ChoiceButtons / SelectedResult を独立コンポーネント化する。
makeAssistantToolUI ラッパー以外は直接テスト可能な状態にし、
helper 8 ケース / Chart 3 ケース / ChoiceButtons 5 ケースの
テストを追加。空配列 colors のフォールバックを併せて改善する。
